CDAP-14224 combine gcs sinks into a single sink
Combining the sinks so that it matches the File sink in
core-plugins. Also refactoring the logic so that it lives
in a file FileFormat class instead of in the BatchSink itself.
This is to prepare for moving the format logic into a common
module that all the file based sources and sinks can depend on.

Further down the line, when formats become actual plugins, the
new FileFormat classes can be re-used.

Also fixing the parent versions to have 5.0.0 as a lower bound.
This is because the plugins use the field level lineage APIs that
were only introduced in CDAP 5.0.0.