CDAP-14338 consolidate terminology and style
Grouped properties into Basic, Advanced, and Credentials sections
with consistent styling for descriptions and placeholders.

Changed any select dropdown for true/false into radio buttons.

Introduced an 'auto-detect' setting for project and credentials
instead of the previous behavior, which expected users to leave
them blank. This is because we found that most users expect to
have to fill out non-advanced properties. In addition, defaulted
the values for those fields to 'auto-detect', as the best user
experience is to run these plugins on dataproc clusters instead
of distributed account keys to every node on a cluster.

Refactored classes to use the same plugin config class hierarchy.

Removed settings from GCS file source for ignoring non existent
input, as it was trying to create some empty directory in a
'tmp' bucket that the user most certainly does not have access to.
Also removing the regex as it didn't seem like it was doing
anything.

Removed useless AbstractFileBatchSource and FileSourceConfig that
were just copied from another project.

Made sure properties show up in the same order in the reference
doc as they do in the plugin widget.

Fixed a bunch of warnings.