CDAP-14442 removed parquet, and orc dependencies
Removed parquet and orc dependencies, as the formats have been
moved to plugins. Did not remove avro since it is used by the
BigQuery plugins.

Removed exports for avro, parquet, and orc. Avro is now a private
dependency, used only within the source and not required by
the program.

Also fixed a NPE that would happen if the big query table does not
exist when hitting the get schema button.