CDAP-15372 fix temporary GCS bucket location
Fixed the BigQuery source and sink to create a temporary GCS
bucket in the same location as the dataset that is being read from
or written to. Without this, the GCS bucket is created in the
default location (US), which cannot write to BigQuery in a different
location.

Also did some refactoring to avoid repeatedly reading the
credentials file, which also makes error handling more clear since
it removes the possibility of throwing IOExceptions from a bunch of
places.