CDAP-15878 remove datastore source schema check
Removed the schema validation check because it is often incorrect,
due to the fact that the source tries to generate a schema based
on a data point that it gets. It is impossible to generate schema
like that, so change the source to just trust the schema given
by the user.