[PLUGIN-178] Allow writing required CDAP schema field to null BQ table field
In #554, we added a validation check that enforced nullability of each CDAP schema field = BQ field.
While this check is correct when CDAP field is nullable and BQ field is required (we shouldn't write
nulls to a required BQ field), the reverse check is incorrect. i.e. we should be able to write a
non-nullable CDAP field to a nullable BQ field because it's ok to write non null values to a nullable
field.