CDAP-19370 fix gcs and bigquery connector tests
Fix the failing gcs connector test. It started to fail due to
changes in the abstract class that it extends. Changed the test
to only check properties set by the GCS connector and not its
base test

Also fixed the BigQuery connector test, which was failing due
to case differences in the entity type (table vs TABLE), and
failing due to sinks and SQL engine being added to the related
plugins.