CDAP-19078 cleanup
cleanup to make things more constants more clear.

fixed service account provider to support configuring the
scopes for the credential, so that it can be used properly
for both bigquery and gcs.

fix places that would only generate auth properties if the
service account is not null to instead always generate the
properties, since we always want to use the custom token provider.