modified the connector tests to pick up information from the env
Modified the connector tests to pick up credentials and project id
from the environment if they are not explicitly given.

Improved the BQ test to create a table and populate it with
known data so that the actual sample content can be verified
instead of just checking that things are not null.

Improved and fixed the Spanner test to create and populate test
data and check actual content.